requirements_prod.txt
pandas==2.2.2
joblib==1.4.2
openpyxl==3.1.2
pydantic==2.7.1
pydantic-settings==2.2.1
sentence-transformers==3.0.1
uvicorn==0.30.1
fastapi==0.111.0
qdrant-client==1.10.1
redis==5.0.7
python-dotenv==1.0.1
sqlalchemy==2.0.33
psycopg2-binary==2.9.9
